# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterDeviceRequest {
  /** @minLength 8 */
  deviceId: string;
  /** @maxLength 60 */
  displayName?: string;
}

export interface PresenceRequest {
  online: boolean;
}

export interface DeviceProfile {
  deviceId: string;
  code: string;
  displayName: string;
  online: boolean;
}

export type ConversationMemberDecision = typeof ConversationMemberDecision[keyof typeof ConversationMemberDecision];


export const ConversationMemberDecision = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface ConversationMember {
  deviceId: string;
  code: string;
  displayName: string;
  decision: ConversationMemberDecision;
}

export interface Message {
  id: string;
  conversationId: string;
  senderDeviceId: string;
  text: string;
  createdAt: string;
  clientId: string;
}

export type ConversationType = typeof ConversationType[keyof typeof ConversationType];


export const ConversationType = {
  direct: 'direct',
  group: 'group',
} as const;

export type ConversationStatus = typeof ConversationStatus[keyof typeof ConversationStatus];


export const ConversationStatus = {
  pending_incoming: 'pending_incoming',
  pending_outgoing: 'pending_outgoing',
  active: 'active',
} as const;

export interface Conversation {
  id: string;
  name: string;
  type: ConversationType;
  status: ConversationStatus;
  members: ConversationMember[];
  lastMessage: Message | null;
}

export interface CreateConversationRequest {
  deviceId: string;
  targetCode: string;
}

export interface CreateGroupRequest {
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @minItems 2 */
  memberCodes: string[];
}

export type DecisionRequestDecision = typeof DecisionRequestDecision[keyof typeof DecisionRequestDecision];


export const DecisionRequestDecision = {
  accept: 'accept',
  reject: 'reject',
} as const;

export interface DecisionRequest {
  deviceId: string;
  decision: DecisionRequestDecision;
}

export interface SendMessageRequest {
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
  clientId: string;
}

export type DeviceIdParameter = string;

export type GetDeviceParams = {
deviceId: DeviceIdParameter;
};

export type UpdateDevicePresenceParams = {
deviceId: DeviceIdParameter;
};

export type ListConversationsParams = {
deviceId: DeviceIdParameter;
};

export type ListMessagesParams = {
deviceId: DeviceIdParameter;
};