# lib/api-zod/src/generated/types/conversationMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMemberDecision } from './conversationMemberDecision';

export interface ConversationMember {
  deviceId: string;
  code: string;
  displayName: string;
  decision: ConversationMemberDecision;
}