# lib/api-zod/src/generated/types/createGroupRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGroupRequest {
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @minItems 2 */
  memberCodes: string[];
}