# lib/api-zod/src/generated/types/decisionRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionRequestDecision } from './decisionRequestDecision';

export interface DecisionRequest {
  deviceId: string;
  decision: DecisionRequestDecision;
}