# lib/api-zod/src/generated/types/registerDeviceRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterDeviceRequest {
  /** @minLength 8 */
  deviceId: string;
  /** @maxLength 60 */
  displayName?: string;
}