# lib/api-zod/src/generated/types/sendMessageRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageRequest {
  deviceId: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
  clientId: string;
}